Fixed "assign while borrowed" not detected in reused specialization
When a specialization is re-used, it now checks any assigns the function does with the current borrows.

In the dirbrowse.lobster example, `entries` was overwritten while a `for` loop had it borrowed, causing a crash. This was possible because the function that does the overwriting had already been typechecked from a context where the variable wasn't borrowed. Now this situation causes the usual "assign while borrowed" error, forcing the programmer to resolve it instead.

Also fixed `switch` being too conservative on how long it borrows its value.